// src/app/SorenTeb/model/Base/ServiceParaclinic/base-service-paraclinic-child-list-model-dto.ts
export class BaseServiceParaclinicChildListModelDto {

  id: number;
  baseServiceId: number;
  serviceName: string;
  baseParaclinicChildId: number;
  paraclinicChildName: string;
  baseCaliforniaID: number;
  californiaCode: string;
  indexQuantity: number;
  indexQuantityAbsence: number;
  indexQuantityTechnical: number;

}